refactor(authentication): Remove LoginResponse and replace with LoginResponseDTO
